+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Render `gitsvg` git-graph diagrams from fenced code blocks to inline SVG at build time.
+
 ### Changed
 
 ### Deprecated